Move http integration tests to the appropriate file
